rename slack utilities for consistency and drop index
