--- src/utils/slack/index.ts
@@ -1,3 +1,0 @@
-export * from './slackWebClient';
-export * from './channel';
-export * from './user';
